Switch to PostgreSQL

+++ support/01-ddl.sql
@@ -11,7 +11,7 @@ create table Products
     Code         int    not null
         primary key,
     Name         text   not null,
-    Price        double not null,
+    Price        float not null,
     Manufacturer int    not null,
     constraint Products_ibfk_1
         foreign key (Manufacturer) references Manufacturers (Code)
